+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bug-fixes within the same version aren't needed
 * Fix regex to detect the pluralized message 'snapshots failed' and show the 'Would you like to update your Snapshots' dialog - shruda
 * Add 'hair space' (U+200A) after decoration text - [@rfgamaral](https://github.com/rfgamaral)
 * Fix decoration color for 'unknown' tests - [@rfgamaral](https://github.com/rfgamaral)
+* Fix remove ANSI characters from test messages - [@jmarceli](https://github.com/jmarceli)
 
 -->